feat(view) add problem view below project explorer (#2307)
Closes STUDIO-3552

+++ bundles/plugins/org.bonitasoft.studio.application/src/org/bonitasoft/studio/application/perspective/WelcomePagePerspectiveFactory.java
@@ -32,7 +32,7 @@ public class WelcomePagePerspectiveFactory extends AbstractPerspectiveFactory {
         final String editorArea = layout.getEditorArea();
         final IFolderLayout leftFolder = layout.createFolder("left", IPageLayout.LEFT, getExplorerViewRatio(), editorArea); //$NON-NLS-1$
         leftFolder.addView(BonitaProjectExplorer.ID);
-        final IFolderLayout rightFolder = layout.createFolder("right", IPageLayout.RIGHT, (float) 0.75, editorArea); //$NON-NLS-1$
+        final IFolderLayout rightFolder = layout.createFolder("right", IPageLayout.RIGHT, 0.75f, editorArea); //$NON-NLS-1$
         rightFolder.addPlaceholder(IIntroConstants.INTRO_VIEW_ID);
     }
